requirements.txt
transformers>=4.44.0
torch>=2.3.0
huggingface_hub>=0.32.0
requests>=2.31.0
llama-cpp-python>=0.2.90
safetensors>=0.4.3
sentencepiece>=0.2.0
accelerate>=0.33.0
python-dotenv>=1.0.1
peft>=0.11.1
lm-eval==0.4.3
matplotlib>=3.8.0
seaborn>=0.13.0
langdetect>=1.0.9
immutabledict>=2.2.5
pandas>=2.2.0
rouge-score>=0.1.2
wandb>=0.17.0
matplotlib
seaborn
# Synthetic data generation
openai>=1.0.0
# Optional: NeMo Data Designer for alternative backend
nemo-microservices[data-designer]>=0.1.0
PyYAML>=6.0

